+++ src/widgets/combo_box.h
@@ -320,7 +320,7 @@ int32_t combo_box_count_options(widget_t* widget);
  *
  * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
  */
-ret_t combo_box_set_selected_index(widget_t* widget, uint32_t index);
+ret_t combo_box_set_selected_index(widget_t* widget, int32_t index);
 
 /**
  * @method combo_box_set_selected_index_by_text